fix(coderd/x/nats): move cluster helper tests
